fix: add redirect to old url for compile docs
The go url points to the old url still, but a redirect is not working. Since `deno compile --help` using the go url, it is linking to a 404.

This was likely broken with #1182, when the go url was changed but a redirect was not added for the new url, missing that the new url also needed a redirect.

Tested locally by visiting the `/runtime/manual/tools/compile` and `/go/compile` urls. 404 before, working after.