Update the docs on Nunjucks and Vento formatting
From the documentation it seemed Nunjucks and Vento formatting were
stable, and this is somewhat reinforced by the output counting them in
saying "Checked [n] files". The output for `deno fmt --help` suffers
from a similar issue, but is missing a bunch of languages from the
introductory descriptions, and does not include Nunjucks or Vento in the
description for the `--unstable-component` flag.